refactor: reorganize note-related components and hooks; remove unused files
