+++ src/app/dashboard/notes/page.js
@@ -1,19 +1,15 @@
 "use client";
 import React, { useState } from "react";
-import useGetAllNotes from "@/hooks/useGetAllNotes";
-import { Card, CardDescription, CardHeader } from "@/components/ui/card";
+import useGetAllNotes from "@/features/notes/hooks/useGetAllNotes";
 import { Button } from "@/components/ui/button";
 import { useRouter } from "next/navigation";
 import Link from "next/link";
-import { useQueryClient } from "@tanstack/react-query";
-import { singleNote } from "@/hooks/useFetchNote";
-import { Skeleton } from "@/components/ui/skeleton";
-import SingleNoteCard from "./components/SingleNoteCard";
+import { LoadingDiv } from "@/features/notes/singleComponents/LoadingDiv";
+import SingleNoteCard from "@/features/notes/SingleNoteCard";
 
 const Page = () => {
-  const { data, isLoading } = useGetAllNotes();
-
   const router = useRouter();
+  const { data, isLoading } = useGetAllNotes();
 
   const onClick = () => {
     router.push("/dashboard/notes/create");
@@ -28,11 +24,7 @@ const Page = () => {
       <div className="flex justify-evenly items-center w-full">
         <h1 className=" font-bold text-lg">Notes - Dashboard</h1>
       </div>
-      {isLoading ? (
-        <p>Loading All Notes...</p>
-      ) : (
-        <NotesDashboard data={data} onClick={onClick} />
-      )}
+      <NotesDashboard data={data} onClick={onClick} />
     </div>
   );
 };
@@ -49,38 +41,6 @@ const NotesDashboard = ({ data, onClick }) => {
         ))}
       </div>
     </div>
-  );
-};
-
-const LoadingDiv = () => {
-  const loadingItems = 5;
-  return (
-    <div className="flex flex-col gap-2 w-full m-2">
-      <div className="flex flex-col justify-evenly items-center w-full">
-        <h1 className="font-bold text-lg">Notes - Dashboard</h1>
-        <Button>Create New Note</Button>
-      </div>
-      <div className="flex gap-2">
-        {Array.from({ length: loadingItems }).map((_, index) => (
-          <LoadingNoteCard key={index} />
-        ))}
-      </div>
-    </div>
-  );
-};
-
-const LoadingNoteCard = () => {
-  return (
-    <Card className="flex flex-col w-[20rem]">
-      <div className="flex flex-col p-2">
-        <CardHeader>
-          <Skeleton className="w-full h-4" />
-        </CardHeader>
-        <CardDescription>
-          <Skeleton className="w-full h-4" />
-        </CardDescription>
-      </div>
-    </Card>
   );
 };
